requirements.txt
streamlit
google-generativeai
pypdf
pdf2image
pytesseract
pandas
python-dotenv
Pillow
